Parser: refactored from recursive descent to scheduled flat stack
Instead of relying on the call stack, transitioned to a fixed stack with a minimal amount of bytes. This should (in theory) save ram in the long run. There is less data lingering in the call stack.

As a result, the ast is written in reverse compared to before. All parse functions write to a specific pointer offset instead of returning their pointer offset. All parse scheduling is now in reverse order.

The code will be harder to wrap one's head around, but it is likely a necessary sacrifice.

Added peek_token() to look ahead.