Codegen: added initial codegen implementation
Codegen reads the AST file to "generate" code. Currently, no code is generated but the groundwork has been laid. Temporarily the codegen will execute the code it would have generated. This is so I can see instant results.

Moved node types and future stack size into their own file so it can be referenced by the parser and the codegen.