Typechecker: added constant expression evaluation phase.
The constant expression evaluation phase will traverse the parse tree for constants. When one is found, it will bubble up it's constant value to the nearest expression or term that has an operator. When both sides of the expression or term have been bubbled up, it will calculate the value and bubble its value up repeating the process.

This means that constants will know which datatype they need to have in order to avoid wrapping.

Added scratch to AST nodes. This is misc data that is used by each stage for internal purposes. Currently used by the typechecker to evaluate constant expressions.

Automatically kill last displayed image.

Made typechecker ensure variables are unique.