Parser: support nested expressions and terms.
Added support for `1 + 1 + 1 + ...` or `1 * 1 * 1 * ...`. Previously you could only have a term that pointed to 2 factors, or an expression that pointed to 2 terms. Now a term can point to a term, and an expression can point to an expression.

Went back to using a future stack for GraphViz generation because GraphViz takes order into account when displaying nodes. After adding nested expressions/terms, the operator would no longer be in the middle, and left and right nodes were no longer obvious.

Added FUTURE_FLAG_INSERT so that expression and term ops know when they need to force insert a new node between their parent and grand parent.