Initial symgen implementation
Symgen is currently used to figure out the sizes of types, including user defined types. I believe this stage will be expanded to generate a symbol table of some sort with function lookups and such. The name will likely change.

Added a 'pending' param to class that will track if it's still trying to figure out all of it's member sizes. Once a member discovers its size, it bubbles that information up to the nearest class and dirties a bit in pending. If a member is parsed that doesn't know its size yet, pending will be incremented and that member will go to the beginning of the stack. If a member then discovers its size, pending will be decremented.

All of this is to say, classes will know how many bytes they take up on the stack. Classes do not have to be declared before they are used.

This stage is very unoptimal for CPU time and disk I/O. Other implementations would require a higher ram cost, so maybe the trade off is worth it?